Updated support for gitops
