+++ integrations/bitbucket-cloud/bitbucket_cloud/gitops/path_validator.py
