+++ integrations/bitbucket-cloud/bitbucket_cloud/webhook/webhook_client.py
@@ -3,7 +3,11 @@ import json
 
 from bitbucket_cloud.client import BitbucketClient
 from loguru import logger
-from bitbucket_cloud.webhook.events import RepositoryEvents, PullRequestEvents, PushEvents
+from bitbucket_cloud.webhook.events import (
+    RepositoryEvents,
+    PullRequestEvents,
+    PushEvents,
+)
 from httpx import HTTPStatusError
 import hashlib
 import hmac
