+++ integrations/bitbucket-cloud/main.py
@@ -15,7 +15,7 @@ from bitbucket_cloud.webhook.processors.pull_request import (
 from bitbucket_cloud.webhook.processors.repository import (
     RepositoryWebhookProcessor,
 )
-from bitbucket_integration.webhook.processors.push import (
+from bitbucket_cloud.webhook.processors.push import (
     PushWebhookProcessor,
 )
 from initialize_client import init_client, init_webhook_client
@@ -23,7 +23,6 @@ from integration import BitbucketFolderResourceConfig, BitbucketFolderSelector
 from bitbucket_cloud.helpers.folder import (
     process_folder_patterns,
 )
-from webhook_processors.push_webhook_processor import PushWebhookProcessor
 
 
 @ocean.on_start()
